Fixed screen capture
Screen capture per se does not seem to work in 10.7.  I've followed the recommendations listed on the Apple site
and now instead use a window that is place in front of everything.  This now appears to be working on both 10.6
and 10.7, and is a bit simpler because we're using a window whether or not there are multiple monitors