More work on LLDisplayWindow
Made changes to handle changes in the display width and height.  Something is still not working because the new window
does not always end up aligned properly with the new display.

Additionally, the LLGabors don't seem to draw once I change to a new OpenGL context.  Need to fix this.